fix formatting of version section
Previously we were getting line breaks at the end of
each line in the source code of this section because
each line ended with two spaces. This made the rendered
document to appear jagged and was probably not intended

What was probably intended was paragraph breaks (or no
breaks at all). I've converted them to paragraph breaks
and it seems to look much better.