Remove HOST metadata
I originally wanted to make sure that the version
number was included in each URL. When I realized that
each section just uses partial URLs I decided that was
probably cleaner, especially since there is no suitable
hostname for us to use.

So the ramification of this change is that each endpoint
will be defined with a partial URL. However, the very first
paragraphs of the API docs show how to construct a full URL
including version number.