Fix network device available types links
The endpoint `/networkDevices/availableTypes` was returning `+ items
(array[ObjectTypeEnum], fixed-type)` as one of the returned attributes.
This is the wrong named declaration.

A `NetworkDeviceAvailableTypeLinks` named declaration has been added
which contains just a `typeUrl` to match the API response payload.
This will replace the `ObjectTypeEnum` named declaration previously
being returned.