Add release notes for 10.1
This pull request is the initial commit for 10.1 release notes. This release
notes file contains all of the additions and changes in the 10.1 release.

This pull request includes section headings to align with industry standards.
Instead of having a separate heading for breaking changes, we have included a
note to highlight the item is a breaking change, and leave the item in the
section appropriate for it.

Standard sections:
- Added
- Changed
- Deprecated
- Removed
- Fixed
- Security