Add release Metasys 10.1, API v2 CHANGELOG
This pull request is the initial commit for 10.1 CHANGELOG. This CHANGELOG
file contains all of the additions and changes in the 10.1 release and past
releases.

This pull request includes section headings to align with industry standards.
Instead of having a separate heading for breaking changes, we have included a
note to highlight the item is a breaking change, and leave the item in the
section appropriate for it.

Standard sections:
- Added
- Changed
- Deprecated
- Removed
- Fixed
- Security