Use `enum values` and remove introduction
Fully qualified enums doesn't necessaily carry the same meaning to everyone
outside of Metasys. To fix this, we're going to change the term to enums
values.

The introduction section was removed from the CHANGELOG being that it isn't
appropriate for a CHANGELOG. The introduction explains what's new in the API
documentation while the CHANGELOG should only discuss changes to the API
product.