Separate API and API documentation notes
The API product and API documentation notes have been separated being that
release notes should be focused on the product, in this case the API.
There is some value in keeping changes to the documentation but mixing them
with product chcnages may confuse developers. To prevent this, the sections
have been seperated.